fix(PYUSDX): downcast earnerRate from uint32 to uint16
Max valid earnerRate is ONE_HUNDRED_PERCENT (10_000), which fits
comfortably in uint16 (max 65_535). The implicit widening to uint32
when calling ContinuousIndexingMath.convertFromBasisPoints is safe.

Also renames MAX_FEE_RATE to MAX_RATE in BaseTest since the constant
bounds both earner rates and fee rates.

+++ test/unit/MultiMintUnit.t.sol
@@ -90,7 +90,7 @@ contract MultiMintTest is BaseTest {
         vm.prank(earnerManager);
         pyusdx.setAccountInfo(address(extension), 500, 0, address(0));
 
-        pyusdx.setAccountRateBps(address(extension), uint24(500));
+        pyusdx.setAccountRateBps(address(extension), uint16(500));
 
         usdc = new MockERC20("USD Coin", "USDC", 6);
         dai = new MockERC20("Dai", "DAI", 18);
